% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00package_help.R
\docType{package}
\name{r4dsnlp-package}
\alias{r4dsnlp-package}
\title{r4dsnlp: Text Analysis of the R4DS Online Learning Community}
\description{
The R4DS Online Learning Community is a community of R learners at all skill
levels working together to improve their skills. The community features a
Slack workspace, on which learners ask questions and mentors help them find
answers. Here we use the BERT NLP model from Google to analyze those
questions and answers.
}
